requirements.txt
dj-database-url==0.3.0
Django==1.8.4
django-filter==0.11.0
django-taggit==0.17.1
djangorestframework==3.2.4
Markdown==2.6.2
psycopg2==2.6.1
